refactor: (#278) mapNotNull → mapNonNull 변경
