+++ simtong-infrastructure/src/main/kotlin/team/comit/simtong/persistence/file/EmployeeCertificatePersistenceAdapter.kt
@@ -3,6 +3,7 @@ package team.comit.simtong.persistence.file
 import org.springframework.stereotype.Component
 import team.comit.simtong.domain.file.model.EmployeeCertificate
 import team.comit.simtong.domain.file.spi.EmployeeCertificatePort
+import team.comit.simtong.global.extension.CollectionExtensionUtils.mapNonNull
 import team.comit.simtong.persistence.file.mapper.EmployeeCertificateMapper
 
 /**
@@ -31,7 +32,7 @@ class EmployeeCertificatePersistenceAdapter(
     ) : List<EmployeeCertificate> {
         return employeeCertificateJpaRepository.saveAll(
             employeeCertificates.map(employeeCertificateMapper::toEntity)
-        ).mapNotNull(employeeCertificateMapper::toDomain)
+        ).mapNonNull(employeeCertificateMapper::toDomain)
     }
 
     override fun queryEmployeeCertificateByNameAndEmployeeNumber(
