+++ simtong-infrastructure/src/main/kotlin/team/comit/simtong/persistence/notification/NotificationPersistenceAdapter.kt
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Component
 import team.comit.simtong.domain.notification.model.Notification
 import team.comit.simtong.domain.notification.model.NotificationReceiver
 import team.comit.simtong.domain.notification.spi.NotificationPort
+import team.comit.simtong.global.extension.CollectionExtensionUtils.mapNonNull
 import team.comit.simtong.persistence.notification.mapper.NotificationMapper
 import team.comit.simtong.persistence.notification.mapper.NotificationReceiverMapper
 import team.comit.simtong.persistence.notification.repository.NotificationJpaRepository
@@ -40,6 +41,6 @@ class NotificationPersistenceAdapter(
     override fun saveAllNotificationReceiver(notificationReceivers: List<NotificationReceiver>): List<NotificationReceiver> {
         return notificationReceiverRepository.saveAll(
             notificationReceivers.map(notificationReceiverMapper::toEntity)
-        ).mapNotNull(notificationReceiverMapper::toDomain)
+        ).mapNonNull(notificationReceiverMapper::toDomain)
     }
 }
