+++ simtong-infrastructure/src/main/kotlin/team/comit/simtong/persistence/schedule/SchedulePersistenceAdapter.kt
@@ -10,7 +10,8 @@ import team.comit.simtong.domain.schedule.model.Schedule
 import team.comit.simtong.domain.schedule.model.Scope
 import team.comit.simtong.domain.schedule.spi.SchedulePort
 import team.comit.simtong.domain.schedule.vo.SpotSchedule
-import team.comit.simtong.persistence.QuerydslExtensionUtils.or
+import team.comit.simtong.global.extension.CollectionExtensionUtils.mapNonNull
+import team.comit.simtong.global.extension.QuerydslExtensionUtils.or
 import team.comit.simtong.persistence.schedule.mapper.ScheduleMapper
 import team.comit.simtong.persistence.schedule.vo.QSpotScheduleVo
 import java.time.LocalDate
@@ -81,7 +82,7 @@ class SchedulePersistenceAdapter(
             )
             .orderBy(schedule.startAt.asc())
             .fetch()
-            .mapNotNull(scheduleMapper::toDomain)
+            .mapNonNull(scheduleMapper::toDomain)
     }
 
     override fun querySchedulesByPeriodAndUserIdAndScope(startAt: LocalDate, endAt: LocalDate, userId: UUID, scope: Scope): List<Schedule> {
@@ -94,7 +95,7 @@ class SchedulePersistenceAdapter(
             )
             .orderBy(schedule.startAt.asc())
             .fetch()
-            .mapNotNull(scheduleMapper::toDomain)
+            .mapNonNull(scheduleMapper::toDomain)
     }
 
     private fun inPeriodScheduleFilter(startAt: LocalDate, endAt: LocalDate) : BooleanExpression {
