+++ simtong-infrastructure/src/main/kotlin/team/comit/simtong/persistence/spot/SpotPersistenceAdapter.kt
@@ -4,6 +4,7 @@ import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Component
 import team.comit.simtong.domain.spot.model.Spot
 import team.comit.simtong.domain.spot.spi.SpotPort
+import team.comit.simtong.global.extension.CollectionExtensionUtils.mapNonNull
 import team.comit.simtong.persistence.spot.mapper.SpotMapper
 import java.util.UUID
 
@@ -27,7 +28,7 @@ class SpotPersistenceAdapter(
 
     override fun queryAllSpot(): List<Spot> {
         return spotJpaRepository.findAll()
-            .mapNotNull(spotMapper::toDomain)
+            .mapNonNull(spotMapper::toDomain)
     }
 
     override fun querySpotByName(name: String): Spot? {
